fix json types until successful build
