+++ src/components/IngredientsAndInstructionsToggle.tsx
@@ -1,11 +1,11 @@
 "use client";
 import { Button, Group, List, Stack, Text } from "@mantine/core";
 import { useState } from "react";
-import renderInstructions from "./RenderedInstructions";
+import RenderedInstructions from "./RenderedInstructions";
 import Link from "next/link";
 
 type IngAndInstToggleProps = {
-  recipe: Recipe;
+  recipe: UserRecipe;
 };
 
 export const IngredientsAndInstructionsToggle = ({
@@ -47,7 +47,7 @@ export const IngredientsAndInstructionsToggle = ({
           })}
         </List>
       )}
-      {view === "instructions" && renderInstructions(recipe)}
+      {view === "instructions" && <RenderedInstructions recipe={recipe} />}
     </Stack>
   );
 };
